Add missing sdk functions (#71)
* Added functions to cover reference model. #65

* More Basics utilities. #65

* Fixed compile and test errors. #65

* Added some missing functions. #65

* Fix typo. #65

+++ morphir/sdk/core/src/morphir/sdk/Basics.scala
@@ -50,7 +50,7 @@ object Basics {
   @inline def greaterThan(a: Int)(b: Int): Bool = a > b
   @inline def greaterThanOrEqual(a: Int)(b: Int): Bool = a >= b
   @inline def min(a: Int)(b: Int): Int = a min b
-  @inline def max(a: Int)(b: Int): Float = a max b
+  @inline def max(a: Int)(b: Int): Int = a max b
   @inline def add(a: Int)(b: Int): Int = a + b
   @inline def subtract(a: Int)(b: Int): Int = a - b
   @inline def multiply(a: Int)(b: Int): Int = a * b
